from Morten Kjeldgaard: Fix MSPEW for possible alignment problems with mode 0 map by using BYTE array. (Can we avoid byte?)

+++ maplib.f
@@ -567,7 +567,7 @@ C     .. Scalar Arguments ..
       INTEGER IUNIT
 C     ..
 C     .. Array Arguments ..
-      REAL X(*)
+      BYTE X(*)
 C     ..
 C     .. Scalars in Common ..
       REAL AMAX,AMEAN,AMIN,ARMS
